+++ components/archive/search/src/main/java/com/flipperdevices/archive/search/viewmodel/SearchViewModel.kt
@@ -2,15 +2,11 @@ package com.flipperdevices.archive.search.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.flipperdevices.archive.api.SearchApi
-import com.flipperdevices.archive.search.api.IS_EXIT_ON_OPEN_KEY
 import com.flipperdevices.archive.search.model.SearchState
 import com.flipperdevices.bridge.dao.api.delegates.KeyParser
 import com.flipperdevices.bridge.dao.api.delegates.key.UtilsKeyApi
-import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
 import com.flipperdevices.bridge.synchronization.api.SynchronizationApi
 import com.flipperdevices.bridge.synchronization.api.SynchronizationState
-import com.github.terrakok.cicerone.Router
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -18,12 +14,9 @@ import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.mapLatest
 import kotlinx.coroutines.launch
-import tangle.inject.TangleParam
 import tangle.viewmodel.VMInject
 
 class SearchViewModel @VMInject constructor(
-    @TangleParam(IS_EXIT_ON_OPEN_KEY)
-    private val exitOnOpen: Boolean,
     private val utilsKeyApi: UtilsKeyApi,
     private val keyParser: KeyParser,
     private val synchronizationApi: SynchronizationApi
@@ -49,13 +42,6 @@ class SearchViewModel @VMInject constructor(
     fun onChangeText(text: String) {
         viewModelScope.launch {
             queryFlow.emit(text)
-        }
-    }
-
-    fun openKeyScreen(router: Router, keyPath: FlipperKeyPath) {
-        router.sendResult(SearchApi.SEARCH_RESULT_KEY, keyPath)
-        if (exitOnOpen) {
-            router.exit()
         }
     }
 }