+++ components/core/navigation/src/main/java/com/flipperdevices/core/navigation/delegates/OnBackPressListener.kt
@@ -1,8 +1,0 @@
-package com.flipperdevices.core.navigation.delegates
-
-interface OnBackPressListener {
-    /**
-     * @return true if we can process it itself
-     */
-    fun onBackPressed(): Boolean
-}