+++ components/core/navigation/src/main/java/com/flipperdevices/core/navigation/delegates/RouterProvider.kt
@@ -1,7 +1,0 @@
-package com.flipperdevices.core.navigation.delegates
-
-import com.github.terrakok.cicerone.Router
-
-interface RouterProvider {
-    val router: Router
-}